\name{benefits}
\alias{benefits}
\docType{data}
\title{benefits}
\description{Fictitious benefits data for employees in a mid-size company}
\usage{data("benefits")}
\format{
  A data frame with 1471 observations on the following 3 variables.
  \describe{
    \item{\code{employee_id}}{Unique identifier for each employee}
    \item{\code{stock_opt_lvl}}{Job level, where 1 = 'Junior' and 5 = 'Senior'}
    \item{\code{trainings}}{Number of trainings completed within the past year}
  }
}
\examples{
data(benefits)
}
\keyword{datasets}
