\name{demographics}
\alias{demographics}
\docType{data}
\title{demographics}
\description{Fictitious demographics data for employees in a mid-size company}
\usage{data("demographics")}
\format{
  A data frame with 1470 observations on the following 7 variables.
  \describe{
    \item{\code{employee_id}}{Unique identifier for each employee}
    \item{\code{age}}{Employee age in years}
    \item{\code{commute_dist}}{Commute distance in miles}
    \item{\code{ed_lvl}}{Education level, where 1 = 'High School', 2 = 'Associate Degree', 3 = 'Bachelor's Degree', 4 = 'Master's Degree', and 5 = 'Doctoral Degree'}
    \item{\code{ed_field}}{Education field associated with most recent degree}
    \item{\code{gender}}{Gender self-identification}
    \item{\code{marital_sts}}{Marital status}
  }
}
\examples{
data(demographics)
}
\keyword{datasets}
