% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.growth.R
\name{hc.growth}
\alias{hc.growth}
\title{Headcount Growth}
\usage{
hc.growth(df, start.date, end.date, start.var, term.var, group.var = NULL)
}
\arguments{
\item{df}{data frame containing start date, end date, and any group variables preferred.}

\item{start.date}{character string containing the queried beginning date.}

\item{end.date}{character string containing the queried ending date.}

\item{start.var}{character string containing name of start date variable.}

\item{term.var}{character string containing name of end date variable.}

\item{group.var}{character string containing the name of a grouping variable. Defaults to NULL.}
}
\description{
Using start date and end date, this function generates the headcount growth percent between two dates.
}
\keyword{growth}
\keyword{headcount}
