% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.R
\name{hc}
\alias{hc}
\title{Headcount}
\usage{
hc(df, start.var, term.var, date, group.var = NULL)
}
\arguments{
\item{df}{data frame containing start date, end date, and any group variables preferred.}

\item{start.var}{character string containing name of start date variable.}

\item{term.var}{character string containing name of end date variable.}

\item{date}{character string of the date to be queried}

\item{group.var}{character string containing the name of a grouping variable. Defaults to NULL.}
}
\description{
Using start date and end date, this function generates the headcount on a specific date.
}
\examples{
hc(df, "Start.date", "End.Date", "2016-12-31", "Department")
}
\keyword{headcount}
