\name{job}
\alias{job}
\docType{data}
\title{job}
\description{Fictitious job data for employees in a mid-size company}
\usage{data("job")}
\format{
  A data frame with 1470 observations on the following 6 variables.
  \describe{
    \item{\code{employee_id}}{Unique identifier for each employee}
    \item{\code{dept}}{Department of which an employee is a member}
    \item{\code{job_lvl}}{Job level, where 1 = 'Junior' and 5 = 'Senior'}
    \item{\code{job_title}}{Job title}
    \item{\code{overtime}}{Flag set to 'Yes' if the employee is nonexempt and works overtime and 'No' if the employee does not work overtime}
    \item{\code{business_travel}}{Business travel frequency}
  }
}
\examples{
data(job)
}
\keyword{datasets}
