\name{performance}
\alias{performance}
\docType{data}
\title{performance}
\description{Fictitious performance data for employees in a mid-size company}
\usage{data("performance")}
\format{
  A data frame with 1470 observations on the following 3 variables.
  \describe{
    \item{\code{employee_id}}{Unique identifier for each employee}
    \item{\code{salary_hike_pct}}{The percent increase in salary for the employee's most recent compensation adjustment (whether due to a standard merit increase, off-cycle adjustment, or promotion)}
    \item{\code{perf_rating}}{Most recent performance rating, where 1 = 'Needs Improvement', 2 = 'Core Contributor', 3 = 'Noteworthy', and 4 = 'Exceptional'}
  }
}
\examples{
data(performance)
}
\keyword{datasets}
