\name{prior_employment}
\alias{prior_employment}
\docType{data}
\title{prior_employment}
\description{Fictitious prior employment data for employees in a mid-size company}
\usage{data("prior_employment")}
\format{
  A data frame with 1470 observations on the following 2 variables.
  \describe{
    \item{\code{employee_id}}{Unique identifier for each employee}
    \item{\code{prior_emplr_cnt}}{Number of prior employers}
  }
}
\examples{
data(prior_employment)
}
\keyword{datasets}
