\name{sentiment}
\alias{sentiment}
\docType{data}
\title{sentiment}
\description{Fictitious sentiment data for employees in a mid-size company}
\usage{data("sentiment")}
\format{
  A data frame with 1470 observations on the following 6 variables.
  \describe{
    \item{\code{employee_id}}{Unique identifier for each employee}
    \item{\code{env_sat}}{Environment satisfaction score measured on a 4-point Likert scale, where 1 = 'Highly Dissatisfied' and 4 = 'Highly Satisfied'}
    \item{\code{engagement}}{Employee engagement score measured on a 4-point Likert scale, where 1 = 'Highly Disengaged' and 4 = 'Highly Engaged'}
    \item{\code{job_sat}}{Job satisfaction score measured on a 4-point Likert scale, where 1 = 'Highly Dissatisfied' and 4 = 'Highly Satisfied'}
    \item{\code{rel_sat}}{Colleague relationship satisfaction score measured on a 4-point Likert scale, where 1 = 'Highly Dissatisfied' and 4 = 'Highly Satisfied'}
    \item{\code{wl_balance}}{Work-life balance score measured on a 4-point Likert scale, where 1 = 'Poor Balance' and 4 = 'Excellent Balance'}
  }
}
\examples{
data(sentiment)
}
\keyword{datasets}
