\name{status}
\alias{status}
\docType{data}
\title{status}
\description{Fictitious data on the active status of employees in a mid-size company}
\usage{data("status")}
\format{
  A data frame with 1470 observations on the following 2 variables.
  \describe{
    \item{\code{employee_id}}{Unique identifier for each employee}
    \item{\code{active}}{Flag set to 'Yes' for active employees and 'No' for inactive employees}
  }
}
\examples{
data(status)
}
\keyword{datasets}
