\name{survey_responses}
\alias{survey_responses}
\docType{data}
\title{survey_responses}
\description{Fictitious survey responses for anonymized employees in a mid-size company}
\usage{data("survey_responses")}
\format{
  A data frame with 400 observations on the following 12 variables.
  \describe{
    \item{\code{belong}}{Belonging score measured on a 5-point Likert scale, where 1 = 'Highly Unfavorable' and 5 = 'Highly Favorable'}
    \item{\code{effort}}{Discretionary Effort score measured on a 5-point Likert scale, where 1 = 'Highly Unfavorable' and 5 = 'Highly Favorable'}
    \item{\code{incl}}{Inclusion score measured on a 5-point Likert scale, where 1 = 'Highly Unfavorable' and 5 = 'Highly Favorable'}
    \item{\code{eng_1}}{Engagement score on item 1 of 3 measured on a 5-point Likert scale, where 1 = 'Highly Disengaged' and 5 = 'Highly Engaged'}
    \item{\code{eng_2}}{Engagement score on item 2 of 3 measured on a 5-point Likert scale, where 1 = 'Highly Disengaged' and 5 = 'Highly Engaged'}
    \item{\code{eng_3}}{Engagement score on item 3 of 3 measured on a 5-point Likert scale, where 1 = 'Highly Disengaged' and 5 = 'Highly Engaged'}
    \item{\code{happ}}{Happiness score measured on a 5-point Likert scale, where 1 = 'Highly Unfavorable' and 5 = 'Highly Favorable'}
    \item{\code{psafety}}{Psychological Safety score measured on a 7-point Likert scale, where 1 = 'Highly Unfavorable' and 7 = 'Highly Favorable'}
    \item{\code{ret_1}}{Retention score on item 1 of 3 measured on a 5-point Likert scale, where 1 = 'Highly Unfavorable' and 5 = 'Highly Favorable'}
    \item{\code{ret_2}}{Retention score on item 2 of 3 measured on a 5-point Likert scale, where 1 = 'Highly Unfavorable' and 5 = 'Highly Favorable'}
    \item{\code{ret_3}}{Retention score on item 3 of 3 measured on a 5-point Likert scale, where 1 = 'Highly Unfavorable' and 5 = 'Highly Favorable'}
    \item{\code{ldrshp}}{Senior Leadership score measured on a 5-point Likert scale, where 1 = 'Highly Unfavorable' and 5 = 'Highly Favorable'}
  }
}
\examples{
data(survey_responses)
}
\keyword{datasets}
