\name{tenure}
\alias{tenure}
\docType{data}
\title{tenure}
\description{Fictitious tenure data for employees in a mid-size company}
\usage{data("tenure")}
\format{
  A data frame with 1470 observations on the following 6 variables.
  \describe{
    \item{\code{employee_id}}{Unique identifier for each employee}
    \item{\code{work_exp}}{Flag set to 'Yes' for active employees and 'No' for inactive employees}
    \item{\code{org_tenure}}{Years at current company}
    \item{\code{job_tenure}}{Years in current job}
    \item{\code{last_promo}}{Years since last promotion}
    \item{\code{mgr_tenure}}{Years under current manager}
  }
}
\examples{
data(tenure)
}
\keyword{datasets}
