% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_namely.R
\name{theme_namely}
\alias{theme_namely}
\title{Namely Theme for ggplot2}
\usage{
theme_namely(base_size = 12, base_family = "Proxima Nova")
}
\description{
Theme function for ggplot2
}
\examples{
theme_namely()
}
\keyword{ggplot2}
\keyword{namely}
\keyword{theme}
